+++ ep/bench/buffer.py
@@ -1,4 +1,1 @@
-"""Thin compatibility shim — all implementation lives in uccl.ep.buffer."""
-from uccl.ep.buffer import Buffer  # noqa: F401
-
-__all__ = ["Buffer"]
+../python/uccl_ep/buffer.py
