+++ ep/bench/test_low_latency.py
@@ -39,7 +39,7 @@ from utils import (
 
 # UCCL import
 try:
-    from uccl.ep import _ep_native as ep
+    from uccl.ep import ep_cpp as ep
 except ImportError as exc:
     import sys
