+++ ep/bench/utils.py
@@ -1,25 +1,1 @@
-"""Thin compatibility shim — all implementation lives in uccl.ep.utils."""
-from uccl.ep.utils import *  # noqa: F401,F403
-from uccl.ep.utils import (  # explicit re-exports for static analysis
-    EventOverlap,
-    EventHandle,
-    bench,
-    bench_kineto,
-    calc_diff,
-    check_nvlink_connections,
-    create_grouped_scores,
-    destroy_uccl,
-    detect_ib_hca,
-    get_cpu_proxies_meta,
-    get_peer_ip,
-    hash_tensor,
-    init_dist,
-    init_dist_under_torchrun,
-    initialize_uccl,
-    inplace_unique,
-    per_token_cast_back,
-    per_token_cast_to_fp8,
-    _fp8_e4m3_dtype,
-    empty_suppress,
-    suppress_stdout_stderr,
-)
+../python/uccl_ep/utils.py
