+++ ep/deep_ep_wrapper/deep_ep/__init__.py
@@ -1,19 +1,25 @@
+import torch
+
 from uccl.ep import (
-    Config,
-    EventHandle,
-    Buffer,
     EventOverlap,
-    check_nvlink_connections,
+    Buffer,
+    Config,
     initialize_uccl,
     destroy_uccl,
+    test_internode,
+    # TODO topk_idx_t
+    buffer, # module
+    utils, # module
+
 )
 
 __all__ = [
     "Config",
-    "EventHandle",
     "Buffer",
     "EventOverlap",
-    "check_nvlink_connections",
     "initialize_uccl",
     "destroy_uccl",
+    "test_internode",
+    "buffer",
+    "utils",
 ]
