--- ep/deep_ep_wrapper/deep_ep/buffer.py
@@ -1,1 +1,0 @@
-../../bench/buffer.py
