--- ep/deep_ep_wrapper/deep_ep/test_internode.py
@@ -1,1 +1,0 @@
-../../bench/test_internode.py
