--- ep/deep_ep_wrapper/deep_ep/utils.py
@@ -1,1 +1,0 @@
-../../bench/utils.py
