+++ ep/python/uccl_ep/.gitignore
@@ -1,1 +1,1 @@
-_ep_native*.so
+ep_cpp*.so
